story_service/requirements.txt
aio_pika==9.5.5
aioredis==2.0.1
python-dotenv==1.1.0
